ps1: [SPU] increase precision of volume input to amplify function
fixes pitch modulation, especially obvious in J-E-N-O-V-A from FF7.

+++ ares/ps1/spu/spu.cpp
@@ -2,7 +2,7 @@
 
 namespace ares::PlayStation {
 
-auto amplify(s32 sample, s16 volume) -> s32 {
+auto amplify(s32 sample, s32 volume) -> s32 {
   return (s64)sample * volume >> 15;
 }
